Change: (org-outline-numbers) Rewrite the monstrosity
This could be slower, since it has no caching and keeps searching up
the buffer, but it seems very fast on this README.org, which is a
fairly large Org file, so it's probably good enough.

I also have another version that uses caching similar to org-ql's.  If
I publish org-ql's caching as a separate package, this could use that.