WIP
It was getting there until I hit a link to a section that GitHub gives
a different HTML ID than Org does (GitHub replaces spaces with dashes,
while Org percent-encodes them).  I don't know if there's a way to fix
this.  Maybe by also changing the anchor-defining functions...which I
already did (earlier in this file)...but which maybe wasn't active in
that file?  I don't know...it's way too complicated.  Shelving this
for now.